Implement orphan object and vote timeouts
